added param_ber. Still defaulting to 0. can be set at runtime e.g. at console by sudo sh -c "echo 100000 > /sys/modules/vspi_drv/parameters/param_ber" Then in avg. each 100.000 out of 2^32 bytes will be modified. But at maximum the current implementation changes just 1 byte in each transfer per direction.

+++ vspi_drv.h
@@ -39,7 +39,7 @@ struct vspi_dev {
 
 // the configurable parameters
 extern int param_major; // major number for char devices
-extern int param_ber; // Bit error rate
+extern unsigned long param_ber; // Bit error rate
 
 // prototypes for shared functions:
 int vspi_open(struct inode *inode, struct file *filep);
